Mob Melee Attack
- Updated mobs melee attack to have a gizmo to show range and to use hitbox detection vs cosmetic attack to wall

+++ Assets/Scripts/Mobs/Mob_Movement.cs
@@ -18,7 +18,7 @@ public class Mob_Movement : MonoBehaviour
     private Mob_Health mobHealth;
 
     //stats
-    [SerializeField] private int moveSpeed;
+    [SerializeField] private float moveSpeed;
